requirements.txt
allure-pytest==2.15.0
pytest==8.4.2
python-dotenv==1.2.1
pytest-retry==1.7.0
pytest-xdist==3.8.0
requests==2.32.5
xdist==0.0.2